DocQABot/requirements.txt
streamlit==1.24.0
nltk==3.8.1
numpy==1.24.3
rank-bm25==0.2.2
python-docx==0.8.11
PyMuPDF==1.22.5 